Deprecate createStyle
